+++ database/database.go
@@ -1,0 +1,34 @@
+package database
+
+import (
+	"context"
+
+	"github.com/cloudquery/cq-provider-sdk/database/postgres"
+	"github.com/cloudquery/cq-provider-sdk/provider/schema"
+	"github.com/hashicorp/go-hclog"
+)
+
+type DB struct {
+	schema.Database
+
+	dialect schema.Dialect
+}
+
+func New(ctx context.Context, logger hclog.Logger, dsn string) (*DB, error) {
+	dType, newDSN, err := DSNtoDialect(dsn)
+	if err != nil {
+		return nil, err
+	}
+
+	dialect := schema.GetDialect(dType)
+
+	db, err := postgres.NewPgDatabase(ctx, logger, newDSN, dialect)
+	if err != nil {
+		return nil, err
+	}
+
+	return &DB{
+		Database: db,
+		dialect:  dialect,
+	}, nil
+}
